[normalization] code refactoring.
